[Enhance] Move script_payload to Script section
